Reading the body from the request spec file
Implement reading of the request body from the request specification
file.

This meant I had to (at least for now) give up on using `Aeson`. It was
a lot simpler to just use the raw string from the spec file directly, so
I decided to do the simpler thing first.