Cookies persisted across requests in the same spec file
Add support for cookies that are persisted across requests in the same
specification file in a similar way as they would be by a web browser.

That is, the `Set-Cookie` header is respected and causes the set
cookie(s) to be sent along with the next request to the same host in the
`Cookie` header.

The heavy lifting of this is done by `Network.HTTP.Client.CookieJar`
from the `http-client` package.

Also added the `-Wall` compilation flag which has helped a lot in
catching mistakes!