Fix bug not allowing paths for HTTP (non-S) requests
Since `fmap` was used on the `Either` the path segments were only added
when the scheme was HTTPS, not HTTP! This is since the `fmap` on
`Either` maps the Right value but keeps the Left value as-is. Changed to
`bimap`, which was what was intended here. It works because `Either` is
a `Bifunctor` instance.