Add support for injecting environment variables into reqeusts
This is very useful for example for handling requests with passwords or
other secrets. Usually you don't want to include those in version
control, and therefore you don't want to write them directly into a
specification file that is checked out in version control. Also the
injected env vars can used for things that change rapidly and therefore
create noise in version control, or for more flexibility of the defined
specification files.