Library version of `surf`
Make a library version of surf in addition to the executable. The
executable now uses the code in the library and just adds the logic to
take a filename argument.

This will make it easier to add tests for `surf` later and makes it
possible for other Haskell libraries/executables to use surf as a
library.